chore: moving folders to their correspondent place
